create a python script in the external directory for HID prox reader
This creates a python script to run on a raspberry pi. The script
is located inside the external directory and is called
prox_reader.py. Currently it it will read a card, query a postgres
server and campare the returned value against the given value. If
they match it will open a relay for 5 seconds. If not it visually
and audibly will notify the user via the HID prox reader of the
invalid code.