add more functionality to application
This PR addes the basic views/models/controller logic for basic
CRUDE for AccessControlDoors and AccessControl. This also creates
the model and associations between AccessControl,
AccessControlDoors, and AccessPermissions. This has CRUDE working
in the first two mentioned modules.

Other Notes:

In this PR I also found an error in many of the controllers in
the edit conditional - render edit. This addresses that fatal
error and fixes it with render 'edit'.